+++ taiji-crypto/src/main/java/com/networknt/taiji/crypto/SignedLedgerEntry.java
@@ -11,7 +11,7 @@ public class SignedLedgerEntry extends LedgerEntry {
 
     private Sign.SignatureData signatureData;
 
-    public SignedLedgerEntry(String to, BigInteger value, String data,
+    public SignedLedgerEntry(String to, long value, String data,
                              Sign.SignatureData signatureData) {
         super(to, value, data);
         this.signatureData = signatureData;
